requirements.txt
# Core ML libraries
tensorflow-macos
tensorflow-metal
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0

# Web development
streamlit>=1.28.0
plotly>=5.17.0
dash>=2.14.0

# Utilities
joblib>=1.3.0
python-dateutil>=2.8.0

# Optional for advanced features
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API development (optional)
fastapi>=0.103.0
uvicorn>=0.23.0
